+++ src/modules/common/dom.js
@@ -35,6 +35,7 @@ export function openModal(modalId, { resetForm = false, formId = null } = {}) {
     if (form) form.reset();
   }
   modal.showModal();
+  document.dispatchEvent(new CustomEvent("formdraftopen", { detail: modalId }));
   return modal;
 }