+++ src/modules/common/feedbackModal.js
@@ -14,6 +14,7 @@
 
 import { notesWithGymNote } from "../../domain/gymNotes.js";
 import { $id, closeModal, openModal, renderMarkupOnce } from "./dom.js";
+import { finishTrainerFormDraft } from "./trainerFormDraft.js";
 
 let deps = null;
 
@@ -37,6 +38,8 @@ export function openFeedbackModal(exId) {
     clientState.exercises[clientState.activeExerciseIndex];
   const client = state.clients.find((c) => c.id === activeClientId);
   const { t } = deps;
+
+  $id("dialog-feedback").dataset.exerciseId = curEx.id;
 
   $id("feedback-client-id").value = activeClientId;
   $id("feedback-exercise-name").value = curEx.name;
@@ -240,6 +243,7 @@ export function setupFeedbackForms() {
             } else {
               currentNoteInput.value = generatedTranscript;
             }
+            currentNoteInput.dispatchEvent(new Event("input", { bubbles: true }));
           }
         }, 1200);
       }
@@ -334,6 +338,7 @@ export function setupFeedbackForms() {
       }
 
       saveToLocalStorage();
+      finishTrainerFormDraft("dialog-feedback");
       renderPendingPlanAdjustments();
       closeModal("dialog-feedback");
     });